create distance.rs in utils crate
